# docs/public-api/references/swagger/process/flowable-swagger-process.yaml
...
swagger: "2.0"
info:
  description: "# flowable / flow??b(??)l /\n\n- a compact and highly efficient workflow\
    \ and Business Process Management (BPM) platform for developers, system admins\
    \ and business users.\n- a lightning fast, tried and tested BPMN 2 process engine\
    \ written in Java. It is Apache 2.0 licensed open source, with a committed community.\n\
    - can run embedded in a Java application, or as a service on a server, a cluster,\
    \ and in the cloud. It integrates perfectly with Spring. With a rich Java and\
    \ REST API, it is the ideal engine for orchestrating human or system activities."
  version: "v1"
  title: "Flowable Process REST API"
  contact:
    name: "Flowable"
    url: "http://www.flowable.org/"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/flowable-rest/service"
...
schemes:
- "http"
- "https"
paths:
...
  /repository/deployments:
    get:
      tags:
      - "Deployment"
      summary: "List Deployments"
      description: ""
      operationId: "listDeployments"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "query"
        description: "Only return deployments with the given name."
        required: false
        type: "string"
      - name: "nameLike"
        in: "query"
        description: "Only return deployments with a name like the given name."
        required: false
        type: "string"
      - name: "category"
        in: "query"
        description: "Only return deployments with the given category."
        required: false
        type: "string"
      - name: "categoryNotEquals"
        in: "query"
        description: "Only return deployments which don?t have the given category."
        required: false
        type: "string"
      - name: "tenantId"
        in: "query"
        description: "Only return deployments with the given tenantId."
        required: false
        type: "string"
      - name: "tenantIdLike"
        in: "query"
        description: "Only return deployments with a tenantId like the given value."
        required: false
        type: "string"
      - name: "withoutTenantId"
        in: "query"
        description: "If true, only returns deployments without a tenantId set. If\
          \ false, the withoutTenantId parameter is ignored."
        required: false
        type: "boolean"
      - name: "sort"
        in: "query"
        description: "Property to sort on, to be used together with the order."
        required: false
        type: "string"
        enum:
        - "id"
        - "name"
        - "deployTime"
        - "tenantId"
      responses:
        200:
          description: "Indicates the request was successful."
          schema:
            $ref: "#/definitions/DataResponseDeploymentResponse"
      security:
      - basicAuth: []
    post:
      tags:
      - "Deployment"
      summary: "Create a new deployment"
      description: "The request body should contain data of type multipart/form-data.\
        \ There should be exactly one file in the request, any additional files will\
        \ be ignored. The deployment name is the name of the file-field passed in.\
        \ If multiple resources need to be deployed in a single deployment, compress\
        \ the resources in a zip and make sure the file-name ends with .bar or .zip.\n\
        \nAn additional parameter (form-field) can be passed in the request body with\
        \ name tenantId. The value of this field will be used as the id of the tenant\
        \ this deployment is done in."
      operationId: "uploadDeployment"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "deploymentKey"
        in: "query"
        required: false
        type: "string"
      - name: "deploymentName"
        in: "query"
        required: false
        type: "string"
      - name: "tenantId"
        in: "query"
        required: false
        type: "string"
      - name: "file"
        in: "formData"
        required: true
        type: "file"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/DeploymentResponse"
        201:
          description: "Indicates the deployment was created."
        400:
          description: "Indicates there was no content present in the request body\
            \ or the content mime-type is not supported for deployment. The status-description\
            \ contains additional information."
      security:
      - basicAuth: []
...
securityDefinitions:
  basicAuth:
    type: "basic"
definitions:
...
  DataResponseDeploymentResponse:
    type: "object"
    properties:
      data:
        type: "array"
        items:
          $ref: "#/definitions/DeploymentResponse"
      total:
        type: "integer"
        format: "int64"
      start:
        type: "integer"
        format: "int32"
      sort:
        type: "string"
      order:
        type: "string"
      size:
        type: "integer"
        format: "int32"
...
  DeploymentResponse:
    type: "object"
    properties:
      id:
        type: "string"
        example: "10"
      name:
        type: "string"
        example: "flowable-examples.bar"
      deploymentTime:
        type: "string"
        format: "date-time"
        example: "2010-10-13T14:54:26.750+02:00"
      category:
        type: "string"
        example: "examples"
      url:
        type: "string"
        example: "http://localhost:8081/flowable-rest/service/repository/deployments/10"
      tenantId:
        type: "string"
...